backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
pypdf
pinecone
langchain-core
langchain-classic
langchain-google-genai
langchain-pinecone
langchain-text-splitters
pydantic
pydantic-settings
python-multipart
duckduckgo-search
ddgs
huggingface_hub
httpx
crewai
crewai-tools
pyjwt
passlib
pytest
pytest-mock
trafilatura
